backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.1
PyYAML==6.0
gunicorn==20.1.0
drf_base64==2.0
gunicorn==20.1.0
django-cors-headers==4.5.0
isort==6.0.1
django-filter==25.1